Fix exim daemon process detection when running under docker. #8
Previously we were identifying the main daemon process by checking to see if an exim process's parent was init (pid 1). This was problematic under docker since exim may be the init process itself, or the child of some other bootstrapping script. Additionally, a child's parent process ID may be changed to init when the parent exits, so this was a bad mechanism in the first place.

We now test if the process group ID matches the process ID itself. This indicates the process leader, and therefore the main exim daemon. This should be more reliable, and has been tested on both Linux and FreeBSD.