Add frozen queue metrics (#47)
Added new `exim_queue_frozen` metric which tracks the number of frozen messages in the queue. This requires peeking at the message headers, which is considerably more expensive compared to simply walking the filesystem for the `exim_queue` metric. There is a new `--queue.read-timeout` option which which prevent the exporter from thrashing on extremely large queues. If the timeout is hit, the exporter increments `exim_queue_read_timeout_errors_total` and stops attempting to read frozen messages until the value of `exim_queue` drops to 20% below the value at the time the timeout occurred. 

---------

Co-authored-by: theS1LV3R <me@s1lv3r.codes>